Move EventType enum into its own file
