+++ VotingApplication/VotingApplication.Data/Model/Event.cs
@@ -3,31 +3,6 @@ using System.ComponentModel.DataAnnotations;
 
 namespace VotingApplication.Data.Model
 {
-    public enum EventType
-    {
-        Error,
-        GoToPage,
-        CreatePoll,
-        UpdateResults,
-        SetExpiry,
-        SetPollType,
-        SetQuestion,
-        SetInviteOnly,
-        SetNamedVoting,
-        SetOptionAdding,
-        SetHiddenResults,
-        AddOption,
-        UpdateOption,
-        DeleteOption,
-        AddVote,
-        DeleteVote,
-        AddBallot,
-        DeleteBallot,
-        Login,
-        Logout,
-        Register
-    }
-
     public class Event
     {
         public Event(EventType eventType, Guid pollId)
