+++ VotingApplication/VotingApplication.Data/VotingApplication.Data.csproj
@@ -121,6 +121,7 @@
     <Compile Include="Migrations\Configuration.cs" />
     <Compile Include="Model\Creation\PollCreationHelper.cs" />
     <Compile Include="Model\Event.cs" />
+    <Compile Include="Model\EventType.cs" />
     <Compile Include="Model\Option.cs" />
     <Compile Include="Model\Poll.cs" />
     <Compile Include="Model\Ballot.cs" />
